Support for global installation.
Also, combined some of the similar functionality.
