+++ controllers/deviceController.js
@@ -1,9 +1,10 @@
 'use strict';
 var inquirer = require('inquirer');
 var q = require('q');
-var OpenT2T = require('opent2t').OpenT2T;
 var helpers = require('../helpers');
 var BaseController = require("./baseController");
+var Opent2tHelper = require("../Opent2tHelper");
+var opent2tHelper = new Opent2tHelper();
 
 class DeviceController extends BaseController {
     constructor() {
@@ -63,7 +64,7 @@ class DeviceController extends BaseController {
         inquirer.prompt(questions).then(function (answers) {
             let prop = state.currentDevice.properties.find(p => p.name === answers.propName);
             let method = 'getDevices' + prop.name.charAt(0).toUpperCase() + prop.name.slice(1);
-            OpenT2T.invokeMethodAsync(state.currentDevice.translator, "", method, [prop.deviceId]).then(info => {
+            opent2tHelper.OpenT2T.invokeMethodAsync(state.currentDevice.translator, "", method, [prop.deviceId]).then(info => {
                 helpers.logObject(info);
                 deferred.resolve(state);
             }).catch(error => {
@@ -97,7 +98,7 @@ class DeviceController extends BaseController {
             let method = 'postDevices' + prop.name.charAt(0).toUpperCase() + prop.name.slice(1);
             try {
                 let parsedValue = JSON.parse(answers.value);
-                OpenT2T.invokeMethodAsync(state.currentDevice.translator, "", method, [prop.deviceId, parsedValue]).then(info => {
+                opent2tHelper.OpenT2T.invokeMethodAsync(state.currentDevice.translator, "", method, [prop.deviceId, parsedValue]).then(info => {
                     helpers.logObject(info);
                     deferred.resolve(state);
                 }).catch(error => {
@@ -131,7 +132,7 @@ class DeviceController extends BaseController {
         inquirer.prompt(questions).then(function (answers) {
             try {
                 let parsedParams = JSON.parse(answers.params);
-                OpenT2T.invokeMethodAsync(state.currentDevice.translator, "", answers.methodName, parsedParams).then(info => {
+                opent2tHelper.OpenT2T.invokeMethodAsync(state.currentDevice.translator, "", answers.methodName, parsedParams).then(info => {
                     helpers.logObject(info);
                     deferred.resolve(state);
                 }).catch(error => {
