+++ index.js
@@ -6,16 +6,15 @@
 
 var program = require('commander');
 var colors = require('colors');
-var OnboardingCli = require("./onboardingCli");
-var TranslatorCli = require("./translatorCli");
+var Opent2tHelper = require("./Opent2tHelper");
 var fs = require('fs');
 var path = require('path');
 var inquirer = require('inquirer');
 var q = require('q');
 var helpers = require('./helpers');
-var translatorCli = new TranslatorCli();
+var opent2tHelper = new Opent2tHelper();
 var MainController = require("./controllers/mainController");
-var configRoot = path.join(__dirname, './');
+var configRoot = process.cwd();
 
 // set theme 
 colors.setTheme({
@@ -56,8 +55,7 @@ else if (program.onboarding) {
 
     let fileName = path.join(configRoot, helpers.createOnboardingFileName(program.hub));
 
-    var onboardingCli = new OnboardingCli();
-    onboardingCli.doOnboarding(program.onboarding).then(info => {
+    opent2tHelper.doOnboarding(program.onboarding).then(info => {
         let configData = helpers.createConfigData(program.hub, program.onboarding, info);
         let data = JSON.stringify(configData);
         helpers.logObject(info);
@@ -87,7 +85,7 @@ else if (program.translator && program.hub) {
     helpers.readFile(fileName, "Please complete onboarding -o").then(data => {
         var configInfo = JSON.parse(data);
         var deviceInfo = configInfo.authInfo;
-        translatorCli.createTranslator(configInfo.translatorPackageName, deviceInfo).then(hub => {
+        opent2tHelper.createTranslator(configInfo.translatorPackageName, deviceInfo).then(hub => {
             var fileName = path.join(configRoot, helpers.createHubDeviceFileName(program.translator, program.id));
             console.log(fileName);
             helpers.readFile(fileName, "Please complete hub -h before calling -t").then(data => {
@@ -102,7 +100,7 @@ else if (program.translator && program.hub) {
                         value = program.di;
                     }
 
-                    translatorCli.getProperty(program.translator, dInfo, program.get, value).then(info => {
+                    opent2tHelper.getProperty(program.translator, dInfo, program.get, value).then(info => {
                         helpers.logObject(info);
                     }).catch(error => {
                         helpers.logError(error);
@@ -117,7 +115,7 @@ else if (program.translator && program.hub) {
                         return;
                     }
 
-                    translatorCli.setProperty(program.translator, dInfo, program.set, program.di, parsedValue).then(info => {
+                    opent2tHelper.setProperty(program.translator, dInfo, program.set, program.di, parsedValue).then(info => {
                         helpers.logObject(info)
                     }).catch(error => {
                         helpers.logError(error);
@@ -137,7 +135,7 @@ else if (program.hub) {
     helpers.readFile(fileName, "Please complete onboarding -o").then(data => {
         var configInfo = JSON.parse(data);
         var deviceInfo = configInfo.authInfo;
-        translatorCli.getProperty(configInfo.translatorPackageName, deviceInfo, 'getPlatforms').then(info => {
+        opent2tHelper.getProperty(configInfo.translatorPackageName, deviceInfo, 'getPlatforms').then(info => {
             helpers.logObject(info);
             helpers.writeArrayToFile(info.platforms, configRoot, "_device_", "controlId");
         }).catch(error => {
@@ -149,13 +147,12 @@ else if (program.hub) {
 }
 else if (program.refreshAuthToken) {
     console.log("------ Refreshing oAuth token for hub %j".header, program.refreshAuthToken);
-    var onboardingCli = new OnboardingCli();
     let fileName = path.join(configRoot, helpers.createOnboardingFileName(program.refreshAuthToken));
     helpers.readFile(fileName, "Please complete onboarding -o").then(data => {
         let configInfo = JSON.parse(data);
         let authInfo = configInfo.authInfo;
-        onboardingCli.loadTranslatorAndGetOnboardingAnswers(configInfo.translatorPackageName).then(answers => {
-            translatorCli.getProperty(configInfo.translatorPackageName, authInfo, 'refreshAuthToken', answers).then(refreshedInfo => {
+        opent2tHelper.loadTranslatorAndGetOnboardingAnswers(configInfo.translatorPackageName).then(answers => {
+            opent2tHelper.getProperty(configInfo.translatorPackageName, authInfo, 'refreshAuthToken', answers).then(refreshedInfo => {
                 helpers.logObject(refreshedInfo);
                 console.log("------ Saving refreshed onboardingInfo to: " + fileName);
                 let configData = helpers.createConfigData(configInfo.translator, configInfo.translatorPackageName, refreshedInfo);
@@ -186,7 +183,7 @@ else if (program.translator) {
         var deviceInfo = JSON.parse(data);
         var dInfo = { 'deviceInfo': deviceInfo, 'hub': undefined };
 
-        translatorCli.getProperty(program.translator, deviceInfo, program.get).then(info => {
+        opent2tHelper.getProperty(program.translator, deviceInfo, program.get).then(info => {
             helpers.logObject(info);
         }).catch(error => {
             helpers.logError(error);
