+++ package.json
@@ -11,11 +11,9 @@
     "commander": "^2.9.0",
     "electron": "^1.6.2",
     "express": "^4.14.0",
-    "findup-sync": "^0.4.3",
     "glob": "^7.1.1",
     "inquirer": "^1.1.3",
     "open": "0.0.5",
-    "opent2t": "^1.0.1",
     "q": "^1.4.1",
     "sanitize-filename": "^1.6.1",
     "xmldom": "^0.1.22",
